first pass at fixing attribute code for our offset-less attributes
git-svn-id: https://jamoma.svn.sourceforge.net/svnroot/jamoma/branches/active@3262 464221e2-0a0f-0410-983b-f8d57d210c17

+++ SourceCode/Framework/RampLib/RampLib.h
@@ -13,7 +13,7 @@
 #include "TTObject.h"
 #include "FunctionLib.h"
 
-typedef void (*RampUnitCallback)(void *, short, double *);
+typedef void (*RampUnitCallback)(void *, long, double *);
 
 
 /****************************************************************************************************/
